service/pyproject.toml
[tool.poetry]
name = "mrdocument"
version = "0.1.0"
description = "PDF OCR and metadata extraction service"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.9.0"
anthropic = "^0.52.0"
pikepdf = "^9.0.0"
pyyaml = "^6.0"
weasyprint = "60.2"
pydyf = "0.8.0"
Pillow = "^10.0.0"
img2pdf = "^0.5.0"
python-docx = "^1.0.0"
reportlab = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.23.0"
pytest-aiohttp = "^1.0.5"

[tool.poetry.scripts]
mrdocument = "mrdocument.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
